Разбивает utils на отдельные компоненты
